sign(x or y) initial on java and put it to response, react take a sign from backend
