+++ src/main/java/org/example/dto/GameStepDTO.java
@@ -7,12 +7,9 @@ public class GameStepDTO {
     private int x;
     @JsonProperty
     private int y;
-    @JsonProperty
-    private String sign;
-    public GameStepDTO(int x, int y, String sign) {
+    public GameStepDTO(int x, int y) {
         this.x = x;
         this.y = y;
-        this.sign = sign;
     }
 
     public GameStepDTO() {
@@ -24,9 +21,5 @@ public class GameStepDTO {
 
     public int getY() {
         return y;
-    }
-
-    public String getSign() {
-        return sign;
     }
 }
